Debugging hero robot...

+++ src/vehicle_system/autoaim/armor_detector/include/armor_tracker/armor_tracker.hpp
@@ -2,7 +2,7 @@
  * @Description: This is a ros-based project!
  * @Author: Liu Biao
  * @Date: 2022-10-13 23:48:07
- * @LastEditTime: 2023-02-02 15:08:05
+ * @LastEditTime: 2023-04-02 18:41:01
  * @FilePath: /TUP-Vision-2023-Based/src/vehicle_system/autoaim/armor_detector/include/armor_tracker/armor_tracker.hpp
  */
 #ifndef ARMOR_TRACKER_HPP_
@@ -47,9 +47,9 @@ namespace armor_detector
         Armor prev_armor;                       //上一次装甲板
         Armor last_armor;                       //本次装甲板
         bool is_initialized;                    //是否完成初始化
-        double last_selected_timestamp = 0.0;            //该Tracker上次被选为目标tracker时间戳
-        double prev_timestamp = 0.0;                     //上次装甲板时间戳
-        double last_timestamp = 0.0;                     //本次装甲板时间戳
+        int64_t last_selected_timestamp = 0.0;            //该Tracker上次被选为目标tracker时间戳
+        int64_t prev_timestamp = 0.0;                     //上次装甲板时间戳
+        int64_t last_timestamp = 0.0;                     //本次装甲板时间戳
         int history_type_sum;                   //历史次数之和
         int selected_cnt = 0;                       //该Tracker被选为目标tracker次数和
         const int max_history_len = 4;          //历史信息队列最大长度
@@ -61,8 +61,8 @@ namespace armor_detector
         std::deque<Armor> history_info_;//目标队列
 
         ArmorTracker();
-        ArmorTracker(Armor src, double src_timestamp);
-        bool update(Armor new_armor, double new_timestamp);
+        ArmorTracker(Armor src, int64_t src_timestamp);
+        bool update(Armor new_armor, int64_t new_timestamp);
         bool calcTargetScore();
     };
 }
